+++ tests/testthat/test-functions.R
@@ -530,6 +530,21 @@ test_that(desc = "Test functions", code = {
   # tima:::.onLoad()
   # tima_full()
 
+  ## Change small params
+  get_example_files()
+  change_params_small()
+  change_params_small(
+    fil_pat = "myExamplePattern",
+    fil_fea_raw = paths$data$source$features,
+    fil_met_raw = paths$data$source$metadata,
+    fil_sir_raw = paths$data$interim$annotations$example_sirius$v6,
+    fil_spe_raw = paths$data$source$spectra,
+    ms_pol = "pos",
+    org_tax = "Gentiana lutea",
+    hig_con = TRUE,
+    summarise = FALSE
+  )
+
   ## CLI arguments check
   arguments <- character()
   arguments$ann_can_fin <- 0